+++ src/public/boot.php
@@ -12,11 +12,12 @@
  *
  * @php-version 7.0
  * @package Core\Public
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
-namespace FireHub\Public;
+namespace FireHub\Core\Public;
+
+/** ### Register the Composer autoloader */
+require __DIR__.'/../../vendor/autoload.php';
 
 require __DIR__ . '/index.php';
 
@@ -26,9 +27,10 @@ use FireHub\Core\Support\Bootstrap\FireHubConfigurator;
  * ### Let there be light
  * @since 1.0.0
  *
- * @throws \FireHub\Core\Throwable\Exception\Bootstrap\FailedToLoadBootloaderException If a bootloader fails to
- * load.
- * @throws \FireHub\Core\Throwable\Exception\Bootstrap\NotBootloaderException If a bootloader is not a bootloader.
+ * @throws \FireHub\Core\Throwable\Exception\Support\Bootstrap\FailedToLoadBootloaderException If a bootloader
+ * fails to load.
+ * @throws \FireHub\Core\Throwable\Exception\Support\Bootstrap\NotBootloaderException If a bootloader is not
+ * a bootloader.
  *
  * @return \FireHub\Core\FireHub
  */