+++ src/public/index.php
@@ -12,11 +12,9 @@
  *
  * @php-version 7.0
  * @package Core\Public
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
-namespace FireHub\Public;
+namespace FireHub\Core\Public;
 
 require __DIR__.'/../support/firehub.LowLevel.php';
 require __DIR__.'/../support/firehub.Autoload.php';