+++ src/shared/constants/firehub.Path.php
@@ -9,8 +9,6 @@
  *
  * @php-version 7.0
  * @package Core\Support
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
 namespace FireHub\Core\Shared\Constants\Path;