+++ src/shared/contracts/magic/firehub.Stringable.php
@@ -9,8 +9,6 @@
  *
  * @php-version 8.0
  * @package Core\Shared
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
 namespace FireHub\Core\Shared\Contracts\Magic;