+++ src/support/bootstrap/firehub.Bootloader.php
@@ -9,8 +9,6 @@
  *
  * @php-version 7.0
  * @package Core
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
 namespace FireHub\Core\Support\Bootstrap;