+++ tests/unit/support/bootstrap/FireHubConfiguratorTest.php
@@ -8,8 +8,6 @@
  * @license https://opensource.org/license/Apache-2-0 Apache License, Version 2.0
  *
  * @package Core\Test
- *
- * @version GIT: $Id$ Blob checksum.
  */
 
 namespace FireHub\Tests\Unit\Support\Bootstrap;
@@ -71,8 +69,8 @@ final class FireHubConfiguratorTest extends Base {
     /**
      * @since 1.0.0
      *
-     * @throws \FireHub\Core\Throwable\Exception\Bootstrap\FailedToLoadBootloaderException
-     * @throws \FireHub\Core\Throwable\Exception\Bootstrap\NotBootloaderException
+     * @throws \FireHub\Core\Throwable\Exception\Support\Bootstrap\FailedToLoadBootloaderException
+     * @throws \FireHub\Core\Throwable\Exception\Support\Bootstrap\NotBootloaderException
      *
      * @return void
      */